Separate the SF-133 load into its own script.
This de-couples the SF-133 load from the process that
loads other domain values that change less frequently.
This is the first step to being able to run SF-133 loads
on demand.
